refactor(academy): move the remaining pure helpers into application/util (AYC-597)
Keeps the folder from being a home for certificate-validity alone while
identical sibling helpers sit loose at the application root. quiz.constants.ts
stays put — *.constants.ts is its own repo-wide convention.
