+++ ayunis-core-backend/src/domain/academy/application/use-cases/reorder-course-modules/reorder-course-modules.use-case.ts
@@ -6,7 +6,7 @@ import {
   ChapterNotFoundError,
   UnexpectedAcademyError,
 } from '../../academy.errors';
-import { assertSameIdSet } from '../../reorder-validation';
+import { assertSameIdSet } from '../../util/reorder-validation';
 import { ReorderCourseModulesCommand } from './reorder-course-modules.command';
 
 @Injectable()
