+++ ayunis-core-backend/src/domain/academy/application/use-cases/update-quiz-question/update-quiz-question.use-case.ts
@@ -6,7 +6,7 @@ import {
   QuizQuestionNotFoundError,
   UnexpectedAcademyError,
 } from '../../academy.errors';
-import { assertValidQuizOptions } from '../../quiz-question-validation';
+import { assertValidQuizOptions } from '../../util/quiz-question-validation';
 import { UpdateQuizQuestionCommand } from './update-quiz-question.command';
 
 @Injectable()
